requirements.txt
fastapi==0.75.0
redis-om==0.0.20
request